dagan/requirements.txt
absl-py==1.0.0
certifi==2021.10.8
cycler==0.11.0
fonttools==4.33.2
grpcio==1.44.0
imageio==2.17.0
importlib-metadata==4.11.3
joblib==1.1.0
kiwisolver==1.4.2
lpips
Markdown==3.3.6
matplotlib==3.5.1
networkx==2.6.3
numpy==1.21.6
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
protobuf==3.20.1
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==5.4.1
scikit-image==0.16.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
tensorboard==1.15.0
threadpoolctl==3.1.0
tifffile==2021.11.2
torch
torchaudio
torchvision
tqdm==4.64.0
typing_extensions==4.2.0
Werkzeug==2.1.1
zipp==3.8.0
